requirements.txt
pyserial==3.5
haversine==2.8.1